refactor: Extension 파일 위치 이동 (#1)
